fix(fsd): align MUS-16 with MUS-34 rebase
- Move app context and hook to shared/public API.

- Move HeaderSearch into widgets layer.

- Fix musiclab.html markup for Prettier.
