+++ src/shared/lib/index.ts
@@ -1,0 +1,9 @@
+export { cn, normalizeString } from "./utils";
+export type { AppContextValue } from "./app-context";
+export {
+  AppContext,
+  defaultArtistOptions,
+  defaultGenreOptions,
+  getSearchQueryFromLocation,
+} from "./app-context";
+export { useAppContext } from "./useAppContext";
