+++ src/shared/lib/useAppContext.ts
@@ -1,0 +1,9 @@
+import { useContext } from "react";
+
+import { AppContext, type AppContextValue } from "./app-context";
+
+export function useAppContext(): AppContextValue {
+  const ctx = useContext(AppContext);
+  if (!ctx) throw new Error("useAppContext must be used within AppProvider");
+  return ctx;
+}
