+++ src/widgets/header-search/index.ts
@@ -1,0 +1,1 @@
+export { HeaderSearch } from "./ui/HeaderSearch";
